% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart.R
\docType{data}
\name{heart}
\alias{heart}
\title{Heart Attack Data}
\format{
A data frame with 4,483 rows and 11 variables:
\describe{
  \item{death}{1 - Death within 48 hours of myocardial infarction onset. 0 - otherwise.}
  \item{anterior}{1 - anterior infarction. 0 - inferior infarction}
  \item{hcabg}{1 - Medical history of cardiac bypass surgery. 0 - otherwise}
  \item{kk1}{1 - Killip class 1. 0 - otherwise}
  \item{kk2}{1 - Killip class 2. 0 - otherwise}
  \item{kk3}{1 - Killip class 3. 0 - otherwise}
  \item{kk4}{1 - Killip class 4. 0 - otherwise}
  \item{age1}{1 - Patient age less than 60. 0 - otherwise}
  \item{age2}{1 - Patient age between 60 and 69 inclusive. 0 - otherwise}
  \item{age3}{1 - Patient age between 70 and 79 inclusive. 0 - otherwise}
  \item{age4}{1 - Patient age 80 or higher. 0 - otherwise}
}
}
\source{
Hardin, James William., and Joseph Hilbe. Generalized Linear Models and Extensions. Stata Press, 2012.

Dataset can be downloaded at \url{https://www.stata-press.com/data/glmext.html}
}
\usage{
data(heart)
}
\description{
A dataset containing 4,483 rows about heart attack victims.
}
\details{
The raw dataset is minimally modified in this package. Rows containing missing values were removed. The center
variable was dropped. This data is referred to as "heart01" in Generalized Linear Models and Extensions book.
}
\keyword{datasets}
